+++ atom/browser/web_contents_preferences.cc
@@ -6,6 +6,7 @@
 
 #include "atom/common/options_switches.h"
 #include "base/command_line.h"
+#include "content/public/common/web_preferences.h"
 
 #if defined(OS_WIN)
 #include "ui/gfx/switches.h"
@@ -72,6 +73,36 @@ void WebContentsPreferences::AppendExtraCommandLineSwitches(
     if (self->web_preferences_.GetBoolean(feature, &b))
       command_line->AppendSwitchASCII(feature, b ? "true" : "false");
   }
+}
+
+void WebContentsPreferences::OverrideWebkitPrefs(
+    content::WebContents* web_contents, content::WebPreferences* prefs) {
+  WebContentsPreferences* self = From(web_contents);
+  CHECK(self);
+
+  bool b;
+  if (self->web_preferences_.GetBoolean("javascript", &b))
+    prefs->javascript_enabled = b;
+  if (self->web_preferences_.GetBoolean("images", &b))
+    prefs->images_enabled = b;
+  if (self->web_preferences_.GetBoolean("java", &b))
+    prefs->java_enabled = b;
+  if (self->web_preferences_.GetBoolean("text-areas-are-resizable", &b))
+    prefs->text_areas_are_resizable = b;
+  if (self->web_preferences_.GetBoolean("webgl", &b))
+    prefs->experimental_webgl_enabled = b;
+  if (self->web_preferences_.GetBoolean("webaudio", &b))
+    prefs->webaudio_enabled = b;
+  if (self->web_preferences_.GetBoolean("web-security", &b)) {
+    prefs->web_security_enabled = b;
+    prefs->allow_displaying_insecure_content = !b;
+    prefs->allow_running_insecure_content = !b;
+  }
+  if (self->web_preferences_.GetBoolean("allow-displaying-insecure-content",
+                                        &b))
+    prefs->allow_displaying_insecure_content = b;
+  if (self->web_preferences_.GetBoolean("allow-running-insecure-content", &b))
+    prefs->allow_running_insecure_content = b;
 }
 
 }  // namespace atom
