+++ atom/browser/web_contents_preferences.h
@@ -12,6 +12,10 @@ namespace base {
 class CommandLine;
 }
 
+namespace content {
+struct WebPreferences;
+}
+
 namespace atom {
 
 // Stores and applies the preferences of WebContents.
@@ -21,9 +25,13 @@ class WebContentsPreferences
   // Get the preferences of |web_contents|.
   static WebContentsPreferences* From(content::WebContents* web_contents);
 
-  // Append command paramters appending to |web_contents|'s preferences.
+  // Append command paramters according to |web_contents|'s preferences.
   static void AppendExtraCommandLineSwitches(
       content::WebContents* web_contents, base::CommandLine* command_line);
+
+  // Modify the WebPreferences according to |web_contents|'s preferences.
+  static void OverrideWebkitPrefs(
+      content::WebContents* web_contents, content::WebPreferences* prefs);
 
   WebContentsPreferences(content::WebContents* web_contents,
                          base::DictionaryValue&& web_preferences);
